Add cargo feature flags for cgp-async
