+++ crates/cgp-async/src/traits/static.rs
@@ -1,0 +1,11 @@
+#[cfg(feature = "static")]
+pub trait MaybeStatic: 'static {}
+
+#[cfg(feature = "static")]
+impl<T: 'static> MaybeStatic for T {}
+
+#[cfg(not(feature = "static"))]
+pub trait MaybeStatic {}
+
+#[cfg(not(feature = "static"))]
+impl<T> MaybeStatic for T {}
