+++ crates/cgp-run/src/lib.rs
@@ -5,7 +5,7 @@ extern crate alloc;
 #[allow(unused_imports)]
 use alloc::boxed::Box;
 
-use cgp_async::{async_trait, Async};
+use cgp_async::{async_trait, Async, MaybeSend};
 use cgp_component::{derive_component, DelegateComponent, HasComponents};
 use cgp_error::HasErrorType;
